+++ distributions/aws/test/e2e/.gitignore
