+++ distributions/aws/test/e2e/resources/custom-resource-templates/katib-experiment-random.yaml
