+++ distributions/aws/test/e2e/utils/constants.py
@@ -1,0 +1,9 @@
+"""
+Global constants module
+"""
+
+DEFAULT_HOST = "http://localhost:8080/"
+DEFAULT_USER_NAMESPACE = "kubeflow-user-example-com"
+DEFAULT_USERNAME = "user@example.com"
+DEFAULT_PASSWORD = "12341234"
+KUBEFLOW_GROUP = "kubeflow.org"
