--- e2e/conftest.py
@@ -1,10 +1,0 @@
-"""
-Add additional pytest supported test configurations.
-
-https://docs.pytest.org/en/6.2.x/example/simple.html
-"""
-
-def pytest_addoption(parser):
-    parser.addoption("--metadata", action="store", help="Metadata file to resume a test class from.")
-    parser.addoption("--keepsuccess", action="store_true", default=False, help="Keep successfully created resources on delete.")
-    parser.addoption("--region", action="store", help="Region to run the tests in. Will be overriden if metadata is provided and region is present.")
