--- e2e/constants.py
@@ -1,9 +1,0 @@
-"""
-Global constants module
-"""
-
-GENERIC_KUSTOMIZE_MANIFEST_PATH = "./manifests/generic"
-DEFAULT_USER_NAMESPACE = "kubeflow-user-example-com"
-
-CUSTOM_RESOURCE_TEMPLATES_FOLDER = "./custom-resource-templates"
-KUBEFLOW_GROUP = "kubeflow.org"
