--- e2e/manifests/generic/kustomization.yaml
@@ -1,55 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-
-resources:
-# Cert-Manager
-- ../../../common/cert-manager/cert-manager/base
-- ../../../common/cert-manager/kubeflow-issuer/base
-# Istio
-- ../../../common/istio-1-9/istio-crds/base
-- ../../../common/istio-1-9/istio-namespace/base
-- ../../../common/istio-1-9/istio-install/base
-# OIDC Authservice
-- ../../../common/oidc-authservice/base
-# Dex
-- ../../../common/dex/overlays/istio
-# KNative
-- ../../../common/knative/knative-serving/overlays/gateways
-- ../../../common/knative/knative-eventing/base
-- ../../../common/istio-1-9/cluster-local-gateway/base
-# Kubeflow namespace
-- ../../../common/kubeflow-namespace/base
-# Kubeflow Roles
-- ../../../common/kubeflow-roles/base
-# Kubeflow Istio Resources
-- ../../../common/istio-1-9/kubeflow-istio-resources/base
-
-
-# Kubeflow Pipelines
-- ../../../apps/pipeline/upstream/env/platform-agnostic-multi-user
-# KFServing
-- ../../../apps/kfserving/upstream/overlays/kubeflow
-# Katib
-- ../../../apps/katib/upstream/installs/katib-with-kubeflow
-# Central Dashboard
-- ../../../apps/centraldashboard/upstream/overlays/istio
-# Admission Webhook
-- ../../../apps/admission-webhook/upstream/overlays/cert-manager
-# Notebook Controller
-- ../../../apps/jupyter/jupyter-web-app/upstream/overlays/istio
-# Jupyter Web App
-- ../../../apps/jupyter/notebook-controller/upstream/overlays/kubeflow
-# Profiles + KFAM
-- ../../../apps/profiles/upstream/overlays/kubeflow
-# Volumes Web App
-- ../../../apps/volumes-web-app/upstream/overlays/istio
-# Tensorboards Web App
--  ../../../apps/tensorboard/tensorboard-controller/upstream/overlays/kubeflow
-# Tensorboard Controller
--  ../../../apps/tensorboard/tensorboards-web-app/upstream/overlays/istio
-# Training Operator
-- ../../../apps/training-operator/upstream/overlays/kubeflow
-# MPI Operator
-- ../../../apps/mpi-job/upstream/overlays/kubeflow
-# User namespace
-- ../../../common/user-namespace/base
